requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
Faker==25.3.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.0
appnope==0.1.3
asgiref==3.8.1
asttokens==2.4.1
black==24.4.1
build==1.0.3
CacheControl==0.13.1
certifi==2022.12.7
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
comm==0.2.0
coverage==7.5.4
crashtest==0.4.1
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.8
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
dulwich==0.21.7
executing==2.0.1
Faker==25.3.0
fastjsonschema==2.19.1
filelock==3.13.1
fsspec==2023.10.0
huggingface-hub==0.17.3
identify==2.5.36
idna==3.4
importlib-metadata==7.0.1
installer==0.7.0
ipykernel==6.26.0
ipython==8.17.2
jaraco.classes==3.3.1
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.5.0
keyring==24.3.0
matplotlib-inline==0.1.6
more-itertools==10.2.0
msgpack==1.0.7
mypy-extensions==1.0.0
mysqlclient==2.2.4
nest-asyncio==1.5.8
nodeenv==1.8.0
numpy==1.26.2
packaging==23.2
parso==0.8.3
pathspec==0.12.1
pexpect==4.8.0
pillow==10.2.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
pre-commit==3.7.0
prompt-toolkit==3.0.40
psutil==5.9.6
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
rapidfuzz==3.6.1
regex==2023.10.3
requests==2.31.0
requests-toolbelt==1.0.0
safetensors==0.4.0
shellingham==1.5.4
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
tokenizers==0.14.1
tomlkit==0.12.3
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers==4.35.0
trove-classifiers==2024.1.31
typing_extensions==4.12.0
urllib3==2.0.7
virtualenv==20.25.0
wcwidth==0.2.10
xattr==0.10.1
zipp==3.17.0
